# secp256k1: Use mod n scalar when signing.